requirements.txt
aiohttp==3.5.4
appnope==0.1.0
appscript==1.1.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
avro==1.9.0
avro-python3==1.9.0
backcall==0.1.0
bayesian-optimization==1.0.1
beautifulsoup4==4.7.1
bleach==3.0.2
blis==0.2.4
boto==2.49.0
boto3==1.9.110
botocore==1.12.110
bz2file==0.98
category-encoders==2.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
colorama==0.4.0
confluent-kafka==1.1.0
cymem==2.0.2
decorator==4.4.0
docutils==0.14
fastavro==0.22.4
Flask==1.1.1
Flask-Session==0.3.1
future==0.17.1
gensim==3.8.0
geocoder==1.38.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kafka==1.3.5
MarkupSafe==1.1.1
multidict==4.5.2
murmurhash==1.0.2
numpy==1.16.2
pandas==0.23.4
patsy==0.5.1
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
python-dateutil==2.7.3
pytz==2018.5
ratelim==0.1.6
requests==2.22.0
reverse-geocoder==1.5.1
s3transfer==0.2.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
smart-open==1.8.4
soupsieve==1.8
spacy==2.1.8
srsly==0.1.0
statsmodels==0.9.0
thinc==7.0.8
tqdm==4.35.0
urllib3==1.24.1
wasabi==0.2.2
webencodings==0.5.1
Werkzeug==0.15.6
yarl==1.3.0
